creating manage.py to have all the functions without click commands and updating .gitignore
