+++ .gitignore
@@ -3,11 +3,8 @@
 
 # Setuptools distribution folder.
 /dist/
+/build/
 
 # Python egg metadata, regenerated from source files by setuptools.
 /*.egg-info
 /*.egg
-
-# Other files
-dist
-build
